+++ cmd/root.go
@@ -17,7 +17,7 @@ import (
 
 var (
 	cfgFile       string
-	version       = "v1.4.1"
+	version       = "v1.5.0"
 	configFileSet bool
 )